Enable static export
