+++ .gitignore
@@ -16,6 +16,7 @@
 # next.js
 /.next/
 /out/
+/web_export/
 
 # production
 /build
