+++ README.md
@@ -34,4 +34,15 @@ You can check out [the Next.js GitHub repository](https://github.com/vercel/next
 The easiest way to deploy your Next.js app is to use the [Vercel Platform](https://vercel.com/new?utm_medium=default-template&filter=next.js&utm_source=create-next-app&utm_campaign=create-next-app-readme) from the creators of Next.js.
 
 Check out our [Next.js deployment documentation](https://nextjs.org/docs/app/building-your-application/deploying) for more details.
+#
+# Static Export
+
+To generate a static version of the site, run:
+
+```bash
+npm run export
+```
+
+The exported files will appear in the `web_export` directory. You can deploy
+this folder to GitHub Pages or any other static hosting provider.
 # registerblock
